# Speed up GC by prefetching during marking