This PR rewrites the core marking loop of the major GC, using prefetching to make better use of the processor's memory parallelism. This removes essentially all of the cache misses that occur during marking, speeding up GC.

On [a microbenchmark with about 800MB of heap](https://gist.github.com/stedolan/4369a0fa27820e27d6e56bee5e412896), the time for a `Gc.full_major` improves from about 1.8 seconds to 0.5 seconds.

Of course, most programs don't spend 100% of their time in GC, so improvements are not generally this dramatic. On the few programs it's been tested on, marking time is reduced by 1/3 - 2/3, leading to overall performance improvements of anywhere around 5-20%, depending on how much GC the program does. (More benchmarking is needed!)

I only expect to see any improvements on programs where the heap is much bigger than the cache. Programs that use less than a few tens of megabytes are unlikely to be improved much by this patch, although they shouldn't get slower.

### Algorithm

The current marking algorithm is a pure depth-first-search: marking always proceeds with the object on top of the mark stack, and new blocks are pushed to the stack as they are found. Since the objects are unlikely to be in cache, the GC spends much of its time stalling on cache misses as it waits for the header of a new object to be loaded.

The algorithm in this PR uses a small (currently 256-entry) circular buffer as a queue in front of the mark stack, known as the _prefetch buffer_. During marking, the next object to be scanned is drawn from the prefetch buffer (if it has at least `Pb_min` elements), or popped from the mark stack if the prefetch buffer is empty or close to empty. When new pointers are discovered during marking, these are prefetched and enqueued in the prefetch buffer rather than being followed immediately. Blocks are only pushed to the mark stack when the prefetch buffer overflows.

This ensures that new objects scanned have generally been prefetched at least `Pb_min` steps ago, which means they are very likely to already be in cache.

### Mark stack changes

This PR also contains a small change to the mark stack structure. It can already contain intervals, to indicate that an object is partially scanned, but those intervals are represented by a `(value, offset)` pair. This patch changes that to a `(start, end)` pair, where `start = value + offset` and `end = value + Wosize_val(value)`. This saves a cache miss in determining `Wosize_val(value)` when traversing the tail end of a long array.

### Configuration parameters

The testing and benchmarking of this patch has been done on Intel x86_64 processors with a non-default configuration:
```
./configure --disable-naked-pointers CC='gcc -Wa,-mbranches-within-32B' AS='as -mbranches-within-32B'
```

The patch should work with naked pointers enabled, although no optimisation work has been done in this configuration and I expect performance to be much worse. 

The `-mbranches-within-32B` option is a workaround for a bug in most recent Intel processors, the [Intel JCC Erratum](https://www.intel.co.uk/content/www/uk/en/support/articles/000055650/processors.html). Intel's microcode workaround for this bug introduces a performance issue in instruction decoding when branch instructions straddle 32-byte boundaries. Most programs are not seriously affected by this as instruction decoding is rarely the bottleneck. However, as the prefetching GC loop generally hits cache, instructions execute quickly (approx 2.9 of them per cycle) and it is possible to become performance-bound by decoding. On the `markbench.ml` microbenchmark above, removing the `-mbranches-within-32B` option had a cost of up to 20%. (The numbers for trunk use the same configuration, although it is not as strongly affected).

(We should perhaps consider enabling `-mbranches-within-32B` by default. It has a 1-2% code size penalty and a performance improvement that's usually irrelevant and occasionally dramatic. It's annoying to ask AMD users to pay the code size penalty for Intel's mistake, though).

### Remaining work

  - **Code cleanup** There's some commented out test / debug code in this patch, and not much commenting otherwise. It needs some cleanup before merging, but this version should be good for initial discussion / testing.
  - **Benchmarking** This has not yet been benchmarked on interestingly-large programs.
  - **Forward_tag optimisation** This patch does not currently implement the Forward_tag lazy short-circuiting optimisation, which is somewhat tricky to fit with the prefetching style. Very preliminary testing has suggested that this optimisation might not be very important, as most lazy values seem to be forced early (and so caught by the _minor_ GC's short-circuiting) or not forced at all. More benchmarking is needed, either to show that this optimisation doesn't matter much or to show that it can be implemented efficiently.

---

_[The design and initial implementation of this patch was in collaboration with Will Hasenplaugh, remaining bugs are my own]_