@xavierleroy 2021-02-06
This sounds extremely promising!  Thank you!

I'm not the most competent reviewer for this kind of code.  Nonetheless I had a look at the diff, and I was put off by lots of commented-out code, and unguarded uses of `__builtin_prefetch` (we now have `caml_prefetch` defined in `caml/misc.h`).  Also, all CI checks are failing.  So, could you please make this PR ready for review and tell us?