@stedolan 2021-02-08
> I'm not the most competent reviewer for this kind of code. Nonetheless I had a look at the diff, and I was put off by lots of commented-out code, and unguarded uses of __builtin_prefetch (we now have caml_prefetch defined in caml/misc.h). Also, all CI checks are failing. So, could you please make this PR ready for review and tell us?

Hmm, there was more of that than I remembered! I've deleted the commented-out code and removed the cause of the CI failures (*), so this should be ready to review now. I ran into #10203 when trying to benchmark, though, so you'll have to wait a bit longer for sandmark numbers.

(*) the issue was an assertion that all zero-length blocks are atoms, which isn't actually true because of how ocamlopt statically allocates empty arrays.