@stedolan 2021-02-15
> When you say that with naked pointers you expect performance to be much worse, I understand that you might mean that you just have not spent time on optimising it yet, but for what it is worth here is a comment about that case.

I mean both that I have not spent time on optimising it, and that I think performance won't be great even with time spent optimising it. The code here often exhausts the available memory parallelism (on the Skylake core I was testing on), so doing more memory accesses is going to take more time, regardless of how well they're prefetched.

Having said that, I think the prefetching strategy you describe is the right one (at least for 64-bit targets, 32-bit ones have a different pagetable).