@gadmm 2021-02-18
I meant for both page tables: for the 32-bit page table, let us hope that the first level is in cache already (the size of the first level is 512 words and it might not be randomly distributed), and only prefetch the second level. I am guessing that this would be good enough. Do you have another strategy in mind?

About exhausting memory parallelism (please correct me if I say anything false, I do not consider myself an expert), I am wondering whether you are already doing something against TLB misses. They can be a bottleneck for memory parallelism, and OCaml programs with default settings are susceptible to them since the runtime will not use huge pages. I tested this on your program without the prefetching optimisation: with default settings (in particular transparent hugepages set to enabled=madvise) I get 1,75% of dTLB load misses, against 0,64% with THP enabled=always, and 0,01% with `OCAMLRUNPARAM=H`. This has a noticeable effect on the gc duration (2.707 s/gc vs. 2.461 s/gc vs. 2.285 s/gc, respectively). Furthermore, it seems that more recent processors tend to offer more memory parallelism than Skylake, but only if your bottleneck is not the TLB.

For reference, THP enabled=always gives a speed-up of about 3% on memory-intensive Coq benchmarks.

(`OCAMLRUNPARAM=H` requires that you pre-allocate 500 huge pages of size 2MB with `hugeadm`, which is likely to fail if your computer was not started recently. A simpler solution is to use jemalloc with the option `thp:always`, which benefits the whole allocated memory and not just the major heap, outperforms the `H` options in terms of TLB misses (0,00%) and still gives good results when the memory is fragmented (0,20% TLB misses vs. having to reboot). For reference see [the command lines I used and the results](https://gist.github.com/gadmm/5782cc5d0010ee734ff5c5f9fbc4e484).)