@stedolan 2021-02-22
Prefetching helps with TLB performance as well. Prefetching allows the processor to handle more memory requests in parallel, which include page walks if necessary.

As you say, hugepages will improve performance due to better TLB utilisation, but I think that's independent of the changes here. Below are some numbers from my laptop, where it seems that hugepages have a smaller effect on the prefetching code than on trunk. (This could be explained by the prefetching code managing to overlap more TLB misses than trunk does) Times are noisy, so only 2 sig figs. Some of the trunk times have even more variance than that.

trunk:
|                     | dTLB misses | time     |
|---------------------|------------:|---------:|
| THP enabled=madvise | 1.1%        | 3.8 s/gc |
| THP enabled=always  | 0.4%        | 3.4 s/gc |
| jemalloc thp:always | 0.2%        | 3.3 s/gc |

prefetching:
|                     | dTLB misses | time      |
|---------------------|------------:|----------:|
| THP enabled=madvise |  0.5% | 0.60 s/gc |
| THP enabled=always  |  0.2% | 0.56 s/gc |
| jemalloc thp:always |  0.02% | 0.56 s/gc |

> Furthermore, it seems that more recent processors tend to offer more memory parallelism than Skylake, but only if your bottleneck is not the TLB.

Interesting! Do you have a reference? (To be specific, the detail I'm most interested in is the number of L1 line fill buffers, as these tend to be the limiting factor on random accesses from a single core. More memory parallelism elsewhere in the system doesn't help GC much, as it can be difficult to saturate using non-sequential accesses from a single core)