@gadmm 2021-02-22
What your figures show corresponds indeed to what I understood with "exhausts the available memory parallelism". I understood it to mean: there is just enough memory parallelism without the page table, but by chance there is nothing spared for prefetching the page table. According to this, even with the strategy I proposed, we should expect it to be noticeably slower with the current page table than without. But given that you did not seem to do anything special for the TLB misses, I expect the combination prefetching+page table+THP to be anywhere between no gain compared to without THP (the TLB was not a bottleneck) to having same performance as without the page table (the TLB was the bottleneck, and prefetching removes the cost of the page table). To test the claim that performance will not be great with the page table, we need to make sure to run numbers with huge pages (which is just a question of using jemalloc with `LD_PRELOAD` as we have seen). Unless I misunderstood and you meant something completely different.

Another implicit question was whether with more memory parallelism you could tweak the parameters to make it go faster, in no-naked-pointers mode, but your answer implicitly seems to say no.

> Interesting! Do you have a reference?

As a non-specialist, I enjoyed Lemire's blog, see for instance [1](https://lemire.me/blog/2019/01/01/memory-level-parallelism-intel-skylake-versus-intel-cannonlake/),[2](https://lemire.me/blog/2019/11/25/memory-parallelism-amd-rome-versus-intel/),[3](https://lemire.me/blog/2018/11/13/memory-level-parallelism-intel-skylake-versus-apple-a12-a12x/) for the claims in question. You can easily find other experiments like this showing increased MLP at L1 level for M1 and Zen2.

Regarding your suggestion of adding `-mbranches-within-32B`, since this is relevant for performance in this patch, how about going ahead and proposing the change (here or in a separate PR?).