@gadmm 2021-08-10
Sure @lpw25 (but it creates more work to open a PR just for this). Let me know @stedolan if you are interested in the improvement/simplification.

For the curious here are some benchmarks. (I offer it without interpretation. @stedolan asked for benchmarks on interestingly-large programs, but for an interpretation it is better to do it with all the details elsewhere.)
- _Before_: OCaml 4.12 plus the fix to work accounting
- _After_: this PR rebased on top of 4.12
- _PT_: with page table
- _NNP_: with `--disable-naked-pointers`
- _After, PT (1)_: what is now in OCaml 4.14
- _After, PT (2)_: the prefetching strategy for the page table that I suggested, which I also tested and which I can propose for review if you let me know you are interested after all.

Things measured:
- pace (ns/w): time spent inside the marking loop to perform one unit of work (i.e. decide that one word is still live this major cycle).
- total time (mm:ss): wall-clock time of opam running (take into account that there is a lot of fluctuations on this measure)
- latency (ms): contribution of the marking loop in a mark slice to the tail latency

The test is to install some ocaml or coq packages with opam, which then runs OCaml (in the first case) or Coq (in the second case). This is with 2 physical cores and 2 jobs; the gains with all logical cores busy are likely to be lesser (since hyperthreading helps keep the physical cores busy during stalls), which would be another interesting benchmark to run.

**OCaml, compiling OCaml, dune, lablgtk3 and couple of smaller packages**

| Throughput | pace (ns/w) | rel. (%) | total time | rel. (%) |
| -------------- | ---------: | ----------: | ----------: | ----------: |
| Before, PT | 18.6 | | 4:09 | |
| Before, NNP)| 18.0 | -3% | 4:08 | -0.4% |
| After, PT (1) | 5.50 | -70% | 3:12 | -22.9% |
| After, PT (2) | 4.86 | -74% | 3:11 | -23.3% |
| After, NNP | 4.16 | -78% | 3:09 | -24.1% |

Latency is not an interesting measure for OCaml itself, and the following table aggregates the measures of different programs so the absolute value is not meaningful in itself (for instance `ocamlc` gives much longer mark slices than `ocamlopt` even though both have the default GC settings); however the relative improvement is consistent across programs and is going to be interesting.

| Latency | max ≤99.99% (ms) | rel. (%) |
| -------------- | ---------: | ----------: |
| Before, PT | 219 | | 
| Before, NNP | 204 | -7% |
| After, PT (1) | 55 | -75% |
| After, PT (2) | 51 | -77% |
| After, NNP | 42 | -81% |

**Coq, compiling some heavy Mathcomp libraries**

(mathcomp-character and mathcomp-odd-order)

| Throughput | pace (ns/w) | rel. (%) | total time | rel. (%) |
| -------------- | ---------: | ----------: | ----------: | ----------: |
| Before, PT |  19.6 | | 13:41 | |
| Before, NNP | 18.8 | -4% | 13:30 | -1.3% |
| After, PT (1) | 7.12 | -64% | 12:55 | -5.6% |
| After, PT (2) | 5.97 | -70% | 12:52 | -6.0% |
| After, NNP | 4.95 | -75% | 12:41 | -7.3% |

Coq's GC settings are optimised for throughput (for instance with a minor heap of 256MB and a space overhead of 200%), so a lesser share of the time is spent in the major GC compared to OCaml. Latency is not an interesting measure for Coq, but the gain is as above.